[cheri_init_globals.h] Avoid using cgetaddr/cgetoffset in assembly
The upcoming standard no longer includes cgetoffset (and cgetaddr has
been removed in ISAv9). For now the assembly code still accepts cgetaddr
and converts it to a move, but use an intrinsic call instead as that can
be optimized to a subregister read on next use rather the requiring an
explicit move.