[CHERIoT] Unify handling of PCC and CGP relocations.
We now treat these as the same and rewrite the instruction for the high
relocation to either AUIPCC or AUICGP depending on whether the target
symbol is in an executable section or not.

This is the first step in refactoring to remove the distinct relocation
types.  This should also make it easy to introduce another kind of
access if we want non-tag-carrying code memory.