[CHERIoT] Zero sret pointers before cross-compartment calls.
Functions that use the sret variant of the calling convention pass a
pointer to space for the callee to store their return values in.  When
this is a cross-compartment call, we need to avoid information leakage.

It is generally better to pass an out parameter and pass a store-only
pointer for this use, but we should still make this work correctly.