[MC] Rework .chericap to support expressions that can fold to sym+const
Currently we require the expression to be sym+const, where the constant
is a known parse-time constant. However, that prevents the use of symbol
difference expressions for the constant, something which LLVM does
itself emit and so cannot round-trip through assembly.

Rework the implementation to look more like nromal data directives. At
parse time, any expression can be used, which will then get folded as
normal, giving an error later if it could not be folded into a
relocatable expression. This also requires taking care to preserve the
provenance within such expressions.