[MC] Support folding capability offset label difference between fragments
In LLVM 15, these expressions would first try and fold the RHS, which
would succeed, and use that as a constant offset for the symbol. In LLVM
17, RISC-V no longer folds label difference expressions within a section
that has instructions due to relaxation, so the RHS remains unfolded,
but cross-term folding will cancel the same symbol if it appears on both
sides with different signs.

Ideally we'd have SET/ADD/SUB-like capability relocations to support
relaxation, but for now just force the RHS folding to happen again for
capability relocations.

Note that when RVC is enabled branches trigger the same code paths as
linker relaxation, even with -mno-relax.