[CHERIoT] Add `cheriot_mmio`, `cheriot_shared_object` and `cheriot_cap_import` attributes
The first two are used from the front-end to indicate that an extern
global declaration refers to a specific compartment import. The third is
the attribute consumed by the back-end to generate the necessary
compartment import entry and the correct sequence of instructions when
taking the address of the global.