[CHERIoT] Fix two issues that block compiling libc++'s extern string templates.
First, clang was crashing in during name mangling when attempting to mangling
Decls with names that are not identifiers. This was only required for checking
for specific fixed names, so we can just bypass it in that case. I have no
idea how to test this, as my manual reductions don't contain an identifier-less
Decl.

Second, add global alias support to RISCVExpandPseudoInsts.