[RISCV] Drop CHERI-RISC-V-specific assembly modifiers and relocations
LLD is now using a normal GOT so all the captable-related names are a
misnomer, instead they should be the same as the normal GOT ones. We can
also drop various other gratuitously-different relocations and modifiers
whilst here. The idea was to make it possible to support mixing integer
and capability mode instructions when it comes to instructions that need
relocations, but that comes down to the same problem that R_RISCV_RELAX
faces when it comes to .option arch, so we shouldn't be doing something
special here that leads to a bunch of unnecessary divergence.

As far as I know we don't have any assembly in our software stack that
uses these modifiers, it all uses the pseudos that hide this detail.