Revert "cheri_init_globals.h: Explicitly set NULL capabilities"
LLD now writes the NULL-derived capability instead of emitting a
capreloc. This code was a bit broken for relative caprelocs, where this
is an offset relative to the load base, not the absolute address, and so
0 could legitimately be a valid object. It also failed to add the
offset, and doesn't work when PCC bounds are correctly set by LLD as
function capabilities can legitimately have their base be the start of
the object even when the function is at a non-zero offset (and it is for
this reason that Morello disables this case downstream, though handled
this case in LLD just as CHERI-MIPS and CHERI-RISC-V used to, so ended
up trying to derive an actual capability, with a non-zero address if a
PIE or DSO).

This reverts commit 18506d32e203b73d52c4dd6ee23f3e385abdd8bf.