[NFC][ELF][CHERI] De-template various capability functions
With a bunch of legacy code gone that was using the provided ELFT we can
stop making these functions templated and remove the need for invokeELFT
in various callers. The MIPS ABI flag is a wart here but CHERI-MIPS
isn't really a thing any more so the slightly gross (and inefficient)
invokeELFT, rather than putting it somewhere in config like Morello LLD
does for its ABI variations, is fine.