[NFCI][ELF][CHERI] Push addCapabilityRelocation into addRelativeReloc
Our use of cap relocs here is equivalent to a relative relocation, so
teach addRelativeReloc how to handle capabilities and align more with
upstream in addGotEntry.

Passing R_CHERI_CAPABILITY rather than R_ABS in the preemptible case
shouldn't matter as I don't think it's read, but is now more consistent.
The intent is that R_ABS is an absolute address and R_CHERI_CAPABILITY
is an absolute capability, and for relative ELF relocs (whether using a
CBuildCap-like or a Morello fragment format) we can add static internal
relocations to write out the capability/fragment bits in the same place
as is done for the addend for Elf_Rel relocations, rather than littering
extra relocation additions throughout.