[NFCI][ELF][CHERI] Move non-zero addend check to addCapabilityRelocation
Passing R_CHERI_CAPABILITY to addReloc here is a hack; it should really
be R_ADDEND since expr here is for the addend not the whole relocation,
but we pass R_CHERI_CAPABILITY to mean "R_ADDEND but it's for a
capability" so we can do this check. Instead, push it to the sole caller
that actually passes R_CHERI_CAPABILITY today (the one that is used for
capability relocations in input files, which is the case we need to
overwrite the addend for on MIPS).