[NFC][ELF][CHERI] Rename useRelativeCheriRelocs to useRelativeElfCheriRelocs
The point of this is to control whether we are using real ELF
relocations to encode what are logically relative relocations, rather
than caprelocs. Whichever is the case, we still need some form of
relative relocation. Make that clearer, as the current name makes some
of the code look very strange (in particular, the calls to the new
addRelativeCapabilityRelocation).