[ELF][Mips][RISCV] Support CHERI_CAPABILITY(_CALL) in relocate
Currently we use symbolicRel as addendRelType for symbolic capability
relocations, but this is an annoying special case to have. Instead,
allow passing the original type through and treat it as a way to write
the addend (or, in future, a word of the raw capability bits, or in the
case of Morello downstream, the fragment).