[NFCI][CodeGen][MC][Mips][RISCV] Follow emitValue for EmitCheriCapability
Rename the underlying function to emitCheriCapability, and have the
symbol-only version be non-virtual and called emitSymbolCheriCapability
like emitSymbolValue.

Drop the addend overloads and let the callers create the adds. Not many
callers exist, no equivalent exists for integers, and two of the three
uses will be changed in a future commit anyway.