[CHERIoT] Fix an issue where large frame indices could cause cause LLVM to generate an offset sequence leading to an invalid pointer.
When the frame is very large, LLVM would try to decompose frame indices by folding the (signed) low 12 bits of the offset into a CIncOffsetImm instruction, materializing the base (containing only the high bits) using an LUI + CIncOffset. This could cause problems when the sign of the low 12 bits did not match the sign of the high bits. In the triggering example, we would generate a sequence which first added a large offset to CSP (taking it out of the representable bounds, clearing the tag) before adding a smaller negative offset to arrive at the intended final offset.

The fix is straightforward: disallow this transformation on CHERIoT when the signs of the two offsets do not match.