[ELF][CHERI] Handle corner case where .text is empty but has a symbol
This is a stupid degenerate case that should never exist in the real
world, but in case that's the input given, we still want to assert the
PCC bounds (which should cover a zero-byte range at the address of the
empty .text, although for now will also include any .rodata-like
sections due to the existing workaround for CheriBSD 25.03's rtld bug).
Otherwise we will emit no PT_CHERI_PCC (which for now would tell the
runtime that it should figure out bounds by itself) and use a base and
length of 0 in relocations.

Whilst touching this function, make it a bit less confusing by adding
some early break and return statements.