[NFC][ELF][CHERI] Move CHERI-MIPS trampoline code into CheriCapRelocsSection
This is a bit weird, but is effectively treated as an implementation
detail of how caprelocs are implemented for this case. If you squint
it's not so dissimilar from how RelocationBaseSection::addReloc adds
static relocations for addends for REL, rather than requiring caller
know to do so (and not provide an addend). Importantly, this lets us
pretend this quirk doesn't exist and use in.capRelocs directly, thus
hiding it deep down in legacy caprelocs-related function without any
extra indirection needed that's not present for non-CHERI/caprelocs.