[NFC][ELF][CHERI] Inline addRelativeCapabilityRelocation
All but one of these uses are places that are only ever for caprelocs,
with the remaining one addRelativeReloc itself, so this will allow
Morello's R_MORELLO_(FUNC_)RELATIVE implementation downstream to be put
there where it belongs.

We can also make use of the overloads and pass references to either
Symbol or InputSectionBase so the level of indirection looks like is
used for the non-caprelocs case.