[CHERI] For global accesses routed via CLLC, attempt to fold the low bits of the dereference into the offset field of the load/store instruction.
This is a bit complicated because of the order of events, and when the necessary information is available. What turns out to be the best approach is to fold CLLC together with eligible dereference operations during the RISCVCheriCleanup pass (pre-RA), which is when we can most easily determine profitability. The concrete expansion has to happen during post-RA pseudo expansion.